+++ packages/kernel/src/kernel.ts
@@ -306,6 +306,20 @@ export class StliteKernel {
     });
   }
 
+  /**
+   * Reboot the Streamlit server.
+   * Note that we also need to refresh (rerender) the frontend app after calling this method
+   * to reflect the changes on the user-facing side.
+   */
+  public reboot(entrypoint: string): Promise<void> {
+    return this._asyncPostMessage({
+      type: "reboot",
+      data: {
+        entrypoint,
+      },
+    });
+  }
+
   private _asyncPostMessage(
     message: InMessage,
   ): Promise<ReplyMessageGeneralReply["data"]>;